+++ tests/test_train.py
@@ -1,6 +1,6 @@
 import unittest
+from unittest.mock import patch, call
 
-from unittest.mock import patch
 from xr2learn_enablers_cli.train import ssl_pipeline
 
 
@@ -11,14 +11,16 @@ class CliTrainTestCase(unittest.TestCase):
     def tearDown(self):
         pass
 
-    # @patch('xr2learn_enablers_cli.train.call_docker')
-    # def test_ssl_pipeline_call_docker_ssl_training_with_correct_parameters_if_not_fe_only(self, mock_call_docker):
-    #     gpu = False
-    #     env_vars = {'env': 'value'}
-    #     modality = 'audio'
-    #     ssl_pre_train = None
-    #     ssl_pipeline(ssl_pre_train, modality, env_vars, gpu)
-    #     mock_call_docker.assert_called_with(ssl_pre_train, modality, env_vars, gpu)
+    @patch('xr2learn_enablers_cli.train.call_docker')
+    def test_ssl_pipeline_call_docker_ssl_training_with_correct_parameters_if_not_fe_only(self, mock_call_docker):
+        gpu = False
+        env_vars = {'env': 'value'}
+        modality = 'audio'
+        ssl_pre_train = 'none'
+        ssl_pipeline(ssl_pre_train, modality, env_vars, gpu)
+        expected_docker_img_name = f'ssl-features-generation-{modality}'
+        mock_call_docker.assert_called_with(expected_docker_img_name, env_vars=env_vars, gpu=gpu)
+        self.assertEqual(mock_call_docker.call_count, 2)
 
     @patch('xr2learn_enablers_cli.train.call_docker')
     def test_ssl_pipeline_call_docker_fe_with_correct_parameters_if_not_encoder_only(self, mock_call_docker):
@@ -26,8 +28,11 @@ class CliTrainTestCase(unittest.TestCase):
         gpu = False
         env_vars = {'env': 'value'}
         modality = 'audio'
-        ssl_pre_train = None
+        # ssl_pre_train is always str type
+        ssl_pre_train = 'none'
         ssl_pipeline(ssl_pre_train, modality, env_vars, gpu)
         expected_docker_img_name = f'ssl-features-generation-{modality}'
-        mock_call_docker.assert_called_with(expected_docker_img_name, env_vars=env_vars, gpu=gpu)
+        calls = [call(expected_docker_img_name, env_vars=env_vars, gpu=gpu), call(expected_docker_img_name, env_vars=env_vars, gpu=gpu)]
+        # mock_call_docker.assert_called_with(expected_docker_img_name, env_vars=env_vars, gpu=gpu)
+        mock_call_docker.assert_has_calls(calls)
         self.assertEqual(mock_call_docker.call_count, 2)